Change NumericToken from IDataToken<decimal> to IDataToken<double> and fix #765

+++ src/UglyToad.PdfPig.Tests/Integration/VisualVerification/SkiaHelpers/SkiaExtensions.cs
@@ -15,7 +15,7 @@
                 0, 0, 1);
         }
 
-        public static SKColor ToSKColor(this IColor pdfColor, decimal alpha)
+        public static SKColor ToSKColor(this IColor pdfColor, double alpha)
         {
             var color = SKColors.Black;
             if (pdfColor != null && pdfColor is not PatternColor)
