+++ src/UglyToad.PdfPig/Graphics/Operations/TextShowing/MoveToNextLineShowTextWithSpacing.cs
@@ -22,12 +22,12 @@
         /// <summary>
         /// The word spacing.
         /// </summary>
-        public decimal WordSpacing { get; }
+        public double WordSpacing { get; }
 
         /// <summary>
         /// The character spacing.
         /// </summary>
-        public decimal CharacterSpacing { get; }
+        public double CharacterSpacing { get; }
 
         /// <summary>
         /// The bytes of the text.
@@ -47,7 +47,7 @@
         /// <param name="wordSpacing">The word spacing.</param>
         /// <param name="characterSpacing">The character spacing.</param>
         /// <param name="text">The text to show.</param>
-        public MoveToNextLineShowTextWithSpacing(decimal wordSpacing, decimal characterSpacing, string text)
+        public MoveToNextLineShowTextWithSpacing(double wordSpacing, double characterSpacing, string text)
         {
             WordSpacing = wordSpacing;
             CharacterSpacing = characterSpacing;
@@ -60,7 +60,7 @@
         /// <param name="wordSpacing">The word spacing.</param>
         /// <param name="characterSpacing">The character spacing.</param>
         /// <param name="hexBytes">The bytes of the text to show.</param>
-        public MoveToNextLineShowTextWithSpacing(decimal wordSpacing, decimal characterSpacing, byte[] hexBytes)
+        public MoveToNextLineShowTextWithSpacing(double wordSpacing, double characterSpacing, byte[] hexBytes)
         {
             WordSpacing = wordSpacing;
             CharacterSpacing = characterSpacing;
@@ -84,9 +84,9 @@
         /// <inheritdoc />
         public void Write(Stream stream)
         {
-            stream.WriteDecimal(WordSpacing);
+            stream.WriteDouble(WordSpacing);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(CharacterSpacing);
+            stream.WriteDouble(CharacterSpacing);
             stream.WriteWhiteSpace();
 
             if (Bytes != null)
