+++ src/UglyToad.PdfPig/Graphics/Operations/TextState/Type3SetGlyphWidth.cs
@@ -2,12 +2,11 @@
 {
     using System.IO;
 
-    /// <inheritdoc />
     /// <summary>
     /// Set width information for the glyph and declare that the glyph description specifies both its shape and its color for a Type 3 font.
     /// wx specifies the horizontal displacement in the glyph coordinate system; 
     /// it must be consistent with the corresponding width in the font's Widths array.
-    /// wy must be 0. 
+    /// wy must be 0.
     /// </summary>
     public class Type3SetGlyphWidth : IGraphicsStateOperation
     {
@@ -22,19 +21,19 @@
         /// <summary>
         /// The horizontal displacement in the glyph coordinate system.
         /// </summary>
-        public decimal HorizontalDisplacement { get; }
+        public double HorizontalDisplacement { get; }
 
         /// <summary>
         /// The vertical displacement in the glyph coordinate system. Must be 0.
         /// </summary>
-        public decimal VerticalDisplacement { get; }
+        public double VerticalDisplacement { get; }
 
         /// <summary>
         /// Create a new <see cref="Type3SetGlyphWidth"/>.
         /// </summary>
         /// <param name="horizontalDisplacement">The horizontal displacement in the glyph coordinate system.</param>
         /// <param name="verticalDisplacement">The vertical displacement in the glyph coordinate system.</param>
-        public Type3SetGlyphWidth(decimal horizontalDisplacement, decimal verticalDisplacement)
+        public Type3SetGlyphWidth(double horizontalDisplacement, double verticalDisplacement)
         {
             HorizontalDisplacement = horizontalDisplacement;
             VerticalDisplacement = verticalDisplacement;
@@ -48,7 +47,7 @@
         /// <inheritdoc />
         public void Write(Stream stream)
         {
-            stream.WriteDecimal(HorizontalDisplacement);
+            stream.WriteDouble(HorizontalDisplacement);
             stream.WriteWhiteSpace();
             stream.WriteNumberText(VerticalDisplacement, Symbol);
         }
