+++ src/UglyToad.PdfPig/PdfFonts/CidFonts/Type0CidFont.cs
@@ -90,7 +90,7 @@
                 return 1000;
             }
 
-            return (double)Descriptor.MissingWidth;
+            return Descriptor.MissingWidth;
         }
 
         public PdfRectangle GetBoundingBox(int characterIdentifier)
