+++ src/UglyToad.PdfPig/PdfFonts/CidFonts/Type2CidFont.cs
@@ -92,7 +92,7 @@
                 return defaultWidth.Value;
             }
 
-            return (double)(Descriptor?.MissingWidth ?? 1000);
+            return Descriptor?.MissingWidth ?? 1000;
         }
 
         public PdfRectangle GetBoundingBox(int characterIdentifier)
